refactor yargs
