+++ src/main.ts
@@ -6,13 +6,9 @@
 // ==LICENSE-END==
 
 import * as debug_ from "debug";
-import { app, dialog } from "electron";
 import * as path from "path";
-import { cli } from "readium-desktop/main/cli/process";
-import { createStoreFromDi, diMainGet } from "readium-desktop/main/di";
-import { winActions } from "readium-desktop/main/redux/actions";
+import { commandLineMainEntry } from "readium-desktop/main/cli";
 import { _PACKAGING, _VSCODE_LAUNCH } from "readium-desktop/preprocessor-directives";
-import { Store } from "redux";
 
 import { setLcpNativePluginPath } from "@r2-lcp-js/parser/epub/lcp";
 import { initGlobalConverters_OPDS } from "@r2-opds-js/opds/init-globals";
@@ -20,10 +16,8 @@ import {
     initGlobalConverters_GENERIC, initGlobalConverters_SHARED,
 } from "@r2-shared-js/init-globals";
 
-import { appActions } from "./main/redux/actions";
-import { RootState } from "./main/redux/states";
-import { initSessions as initSessionsNoHTTP } from "./main/streamerNoHttp";
-import { tryCatch } from "./utils/tryCatch";
+import { initSessions as initSessionsNoHTTP } from "./main/streamer/streamerNoHttp";
+import { start } from "./main/start";
 
 // import { initSessions as initSessionsHTTP } from "@r2-navigator-js/electron/main/sessions";
 
@@ -71,57 +65,11 @@ setLcpNativePluginPath(lcpNativePluginPath);
 // }
 initSessionsNoHTTP();
 
-export const main = async (flushSession = false): Promise<Store<RootState>> => {
-
-    debug("main fct");
-
-    let store = await tryCatch(() => diMainGet("store"), "Store not init");
-    if (store) {
-        return store;
-    }
-
-    store = await createStoreFromDi();
-    debug("store loaded");
-
-    try {
-
-        if (flushSession) {
-
-            const readers = store.getState().win.session.reader;
-            for (const key in readers) {
-                if (readers[key]) {
-
-                    const reader = readers[key];
-                    store.dispatch(winActions.session.unregisterReader.build(reader.identifier));
-                    store.dispatch(winActions.registry.registerReaderPublication.build(
-                        reader.publicationIdentifier,
-                        reader.windowBound,
-                        reader.reduxState,
-                    ));
-                }
-            }
-        }
-
-        store.dispatch(appActions.initRequest.build());
-        debug("STORE MOUNTED -> MOUNTING THE APP NOW");
-
-    } catch (err) {
-        const message = `REDUX STATE MANAGER CAN'T BE INITIALIZED, ERROR: ${JSON.stringify(err)} \n\nYou should remove your 'AppData' folder\nThorium Exit code 1`;
-        process.stderr.write(message);
-
-        dialog.showErrorBox("THORIUM ERROR", message);
-
-        app.exit(1);
-    }
-
-    return store;
-};
-
 if (_VSCODE_LAUNCH === "true") {
     // tslint:disable-next-line: no-floating-promises
-    main();
+    start();
 } else {
-    cli(); // call main fct
+    commandLineMainEntry(); // call main fct
 }
 
 debug("Process version:", process.versions);
