+++ src/main/redux/sagas/api/publication/import/importPublicationFromFs.ts
@@ -12,7 +12,7 @@ import { acceptedExtensionObject } from "readium-desktop/common/extension";
 import { lcpLicenseIsNotWellFormed } from "readium-desktop/common/lcp";
 import { RandomCustomCovers } from "readium-desktop/common/models/custom-cover";
 import { convertMultiLangStringToString } from "readium-desktop/main/converter/tools/localisation";
-import { extractCrc32OnZip } from "readium-desktop/main/crc";
+import { extractCrc32OnZip } from "readium-desktop/main/tools/crc";
 import {
     PublicationDocument, PublicationDocumentWithoutTimestampable,
 } from "readium-desktop/main/db/document/publication";
