+++ src/main/redux/sagas/api/publication/search.ts
@@ -8,7 +8,7 @@
 import { call as callTyped } from "typed-redux-saga/macro";
 import { PublicationView } from "readium-desktop/common/views/publication";
 import { diMainGet } from "readium-desktop/main/di";
-import { aboutFiltered } from "readium-desktop/main/filter";
+import { aboutFiltered } from "readium-desktop/main/tools/filter";
 import { SagaGenerator } from "typed-redux-saga";
 import { PublicationDocument } from "readium-desktop/main/db/document/publication";
 import { PublicationViewConverter } from "readium-desktop/main/converter/publication";
