+++ src/main/services/lcp.ts
@@ -43,7 +43,7 @@ import { Publication as R2Publication } from "@r2-shared-js/models/publication";
 
 import { injectBufferInZip } from "@r2-utils-js/_utils/zip/zipInjector";
 
-import { extractCrc32OnZip } from "../crc";
+import { extractCrc32OnZip } from "../tools/crc";
 import { lcpActions } from "../redux/actions";
 import { streamerCachedPublication } from "../streamerNoHttp";
 import { DeviceIdManager } from "./device";
