+++ src/main/streamer/streamerCommon.ts
