+++ src/main/streamer/streamerNoHttp.ts
