+++ src/main/tools/crc.ts
